Refactor on attack class, Part 1
This started the refactor on the attack class. Instead of trying to make
the avgval method perform 100 attacks, a separate named perform was
created. This method will perform 1 attack. The avgval method will later
be remade to just run the perform method over and over and summarize the
results. This will improve how the class works.