Rollattack functional
This code is a horrendous mess, I cant even express how many variables
have "hearts" vs "heart" or att_mods vs att_mod. The sheer number of
inconsistencies in pluralization is insane. This code needs to be
reviewed and rewritten badly.

Still, now the rollattack method is functional, and works for cases
where there is a surge directive and where there isn't. Haven't tested
whether ranged surge directives work or not. Still a lot of checking,
modifying, and improving the code to go.... but it's progress.

I've been getting this weird matlab crash occasionally, not sure if it's
related to the code or I just need to restart. I'll commit this and
restart. Here's hoping that solves it.