Currently broken
Added some bug fixes. That matlab crash persists, but only if you click
the run button on the test script too quickly.

Code is currently broken, on the line where it tries to find objects
that only modify range. To be fixed another day.

Tested plenty of cases with insufficient range; all seems to work fine.